[chore] move tests to non deprecated packages, fix chloggen
