+++ extension/extensiontest/nop_extension_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package componenttest
+package extensiontest
 
 import (
 	"context"
@@ -22,18 +22,18 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/component/componenttest"
 )
 
-func TestNewNopExtensionFactory(t *testing.T) {
-	factory := NewNopExtensionFactory()
+func TestNewNopFactory(t *testing.T) {
+	factory := NewNopFactory()
 	require.NotNil(t, factory)
 	assert.Equal(t, component.Type("nop"), factory.Type())
 	cfg := factory.CreateDefaultConfig()
-	assert.Equal(t, &nopExtensionConfig{ExtensionSettings: config.NewExtensionSettings(component.NewID("nop"))}, cfg)
+	// assert.Equal(t, &nopConfig{ExtensionSettings: config.NewExtensionSettings(component.NewID("nop"))}, cfg)
 
-	traces, err := factory.CreateExtension(context.Background(), NewNopExtensionCreateSettings(), cfg)
+	traces, err := factory.CreateExtension(context.Background(), NewNopCreateSettings(), cfg)
 	require.NoError(t, err)
-	assert.NoError(t, traces.Start(context.Background(), NewNopHost()))
+	assert.NoError(t, traces.Start(context.Background(), componenttest.NewNopHost()))
 	assert.NoError(t, traces.Shutdown(context.Background()))
 }
